Add module for more Secure Boot convenience
This uses Ubuntu's bootloader to boot into Ubuntu (if you have one on
your drive) and Windows 10 Recovery. It is still possible to chainload
into the real boot menu, which may require hash enrollment if running in
a Secure Boot environment. When no Secure Boot is present (this is
detected by loading and unloading an unsigned module), directly boot
into the normal boot menu.

This is still considered experimental and therefore available as a
module and not as part of the main download.